# server/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "moncamp 수집·집계 서버",
    "description": "GA4 가 돌려주지 않는 **원본 이벤트**를 내 DB 에 남기는 서버입니다.\n\n- 화면(moncamp.kr)은 이 서버를 **런타임에 읽지 않습니다** — 보내기만 하고, 받아 오는 것은 빌드입니다.\n- `POST /v1/events` 는 답을 기다리지 않는 자리라 `204` 만 돌려줍니다.\n- IP 는 어디에도 저장하지 않습니다. 분당 한도가 받는 순간 보고 버릴 뿐입니다.\n- 원본은 **12개월** 뒤 지우고 날짜·이름별 합계만 남깁니다.\n\n설계 배경: [개발 문서 §2.28](https://github.com/MinsangKwak/pogo-rank/blob/main/docs/DEVELOPMENT.md) ·\n[server/README.md](https://github.com/MinsangKwak/pogo-rank/blob/main/server/README.md) ·\n[server/SCHEMA.md](https://github.com/MinsangKwak/pogo-rank/blob/main/server/SCHEMA.md)",
    "version": "1.0.0",
    "license": {
      "name": "MIT",
      "url": "https://github.com/MinsangKwak/pogo-rank/blob/main/LICENSE"
    }
  },
  "components": {
    "securitySchemes": {
      "adminToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "`Authorization: Bearer <ADMIN_TOKEN>`. 운영에서는 32자 이상이어야 서버가 뜬다"
      },
      "accessToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "`Authorization: Bearer <액세스 토큰>`. `POST /v1/auth/refresh` 가 준다"
      }
    },
...
  },
  "paths": {
    "/v1/auth/google/start": {
      "get": {
        "summary": "구글 로그인 화면으로 보낸다",
        "tags": [
          "인증"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "next",
            "required": false,
            "description": "마치고 돌아갈 앱 안의 경로"
          }
        ],
        "responses": {
          "302": {
            "description": "구글로"
          }
        }
      }
    },
    "/v1/auth/google/callback": {
      "get": {
        "summary": "구글이 인가 코드를 돌려주는 자리",
        "tags": [
          "인증"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "code",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "state",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "error",
            "required": false,
            "description": "사용자가 구글 화면에서 취소하면 온다"
          }
        ],
        "responses": {
          "302": {
            "description": "앱으로"
          }
        }
      }
    },
    "/v1/auth/refresh": {
      "post": {
        "summary": "리프레시 쿠키로 액세스 토큰을 받는다",
        "tags": [
          "인증"
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "access": {
                      "type": "string",
                      "description": "액세스 토큰. 메모리에만 두고 저장소에 안 담는다"
                    },
                    "expiresIn": {
                      "type": "integer",
                      "description": "초"
                    },
                    "role": {
                      "type": "string",
                      "enum": [
                        "pending",
                        "approved",
                        "admin",
                        "root"
                      ]
                    },
                    "beta": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "access",
                    "expiresIn",
                    "role",
                    "beta"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "reason": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/auth/logout-all": {
      "post": {
        "summary": "모든 기기의 로그인을 끊는다",
        "tags": [
          "인증"
        ],
        "security": [
          {
            "accessToken": []
          }
        ],
        "responses": {
          "204": {
            "description": "끊었다"
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "reason": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/sessions": {
      "get": {
        "summary": "어디서 로그인했는지 — IP 는 담지 않는다 (CLAUDE.md §3)",
        "tags": [
          "인증"
        ],
        "security": [
          {
            "accessToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "sessions": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "userAgent": {
                            "type": "string"
                          },
                          "issuedAt": {
                            "type": "string"
                          },
                          "expiresAt": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "userAgent",
                          "issuedAt",
                          "expiresAt"
                        ]
                      }
                    }
                  },
                  "required": [
                    "sessions"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "reason": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/v1/me": {
      "get": {
        "summary": "내 정보",
        "tags": [
          "인증"
        ],
        "security": [
          {
            "accessToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "email": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "picture": {
                      "type": "string"
                    },
                    "role": {
                      "type": "string",
                      "enum": [
                        "pending",
                        "approved",
                        "admin",
                        "root"
                      ]
                    },
                    "beta": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "id",
                    "email",
                    "name",
                    "picture",
                    "role",
                    "beta"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "reason": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "reason": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "계정을 지운다 — 세션과 담아 둔 것이 같이 사라진다",
        "tags": [
          "인증"
        ],
        "security": [
          {
            "accessToken": []
          }
        ],
        "responses": {
          "204": {
            "description": "지웠다"
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "reason": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "https://api.moncamp.kr",
      "description": "운영"
    },
    {
      "url": "http://localhost:8080",
      "description": "로컬"
    }
  ],
...
}